+++ GoogleSignIn/Tests/Unit/GIDSignInTest.m
@@ -16,11 +16,8 @@
 #import <UIKit/UIKit.h>
 #import <XCTest/XCTest.h>
 
-#import "GoogleSignIn/Sources/Public/GoogleSignIn/GIDAuthentication.h"
-#import "GoogleSignIn/Sources/Public/GoogleSignIn/GIDConfiguration.h"
-#import "GoogleSignIn/Sources/Public/GoogleSignIn/GIDGoogleUser.h"
-#import "GoogleSignIn/Sources/Public/GoogleSignIn/GIDProfileData.h"
-#import "GoogleSignIn/Sources/Public/GoogleSignIn/GIDSignIn.h"
+// Test module imports
+@import GoogleSignIn;
 
 #import "GoogleSignIn/Sources/GIDGoogleUser_Private.h"
 #import "GoogleSignIn/Sources/GIDSignInInternalOptions.h"
@@ -136,6 +133,9 @@ static NSString *const kTokenURL = @"https://oauth2.googleapis.com/token";
 static NSString *const kFakeURL = @"http://foo.com";
 
 static NSString *const kEMMSupport = @"1";
+
+static NSString *const kGrantedScope = @"grantedScope";
+static NSString *const kNewScope = @"newScope";
 
 /// Unique pointer value for KVO tests.
 static void *kTestObserverContext = &kTestObserverContext;
@@ -315,10 +315,6 @@ static void *kTestObserverContext = &kTestObserverContext;
   };
 
   [_signIn addObserver:self
-            forKeyPath:NSStringFromSelector(@selector(clientID))
-               options:0
-               context:kTestObserverContext];
-  [_signIn addObserver:self
             forKeyPath:NSStringFromSelector(@selector(currentUser))
                options:0
                context:kTestObserverContext];
@@ -338,9 +334,6 @@ static void *kTestObserverContext = &kTestObserverContext;
   [super tearDown];
 
   [_signIn removeObserver:self
-               forKeyPath:NSStringFromSelector(@selector(clientID))
-                  context:kTestObserverContext];
-  [_signIn removeObserver:self
                forKeyPath:NSStringFromSelector(@selector(currentUser))
                   context:kTestObserverContext];
 }
@@ -353,7 +346,7 @@ static void *kTestObserverContext = &kTestObserverContext;
   XCTAssertTrue(signIn1 == signIn2, @"shared instance must be singleton");
 }
 
-- (void)testRestoredGoogleUserFromPreviousSignIn_hasPreviousUser {
+- (void)testRestorePreviousSignInNoRefresh_hasPreviousUser {
   [[[_authorization expect] andReturn:_authState] authState];
   OCMStub([_authState lastTokenResponse]).andReturn(_tokenResponse);
   OCMStub([_tokenResponse scope]).andReturn(nil);
@@ -367,21 +360,23 @@ static void *kTestObserverContext = &kTestObserverContext;
   OCMStub([idTokenDecoded initWithIDTokenString:OCMOCK_ANY]).andReturn(idTokenDecoded);
   OCMStub([idTokenDecoded subject]).andReturn(kFakeGaiaID);
 
-  GIDGoogleUser *previousUser = [_signIn restoredGoogleUserFromPreviousSignIn];
+  [_signIn restorePreviousSignInNoRefresh];
 
   [_authorization verify];
   [_authState verify];
   [_tokenResponse verify];
-  XCTAssertEqual(previousUser.userID, kFakeGaiaID);
+  XCTAssertEqual(_signIn.currentUser.userID, kFakeGaiaID);
+
+  [idTokenDecoded stopMocking];
 }
 
-- (void)testRestoredGoogleUserFromPreviousSignIn_hasNoPreviousUser {
+- (void)testRestoredPreviousSignInNoRefresh_hasNoPreviousUser {
   [[[_authorization expect] andReturn:nil] authState];
 
-  GIDGoogleUser *previousUser = [_signIn restoredGoogleUserFromPreviousSignIn];
+  [_signIn restorePreviousSignInNoRefresh];
 
   [_authorization verify];
-  XCTAssertNil(previousUser);
+  XCTAssertNil(_signIn.currentUser);
 }
 
 - (void)testHasPreviousSignIn_HasBeenAuthenticated {
@@ -461,6 +456,67 @@ static void *kTestObserverContext = &kTestObserverContext;
                restoredSignIn:YES
                oldAccessToken:YES
                   modalCancel:NO];
+}
+
+- (void)testAddScopes {
+  // Restore the previous sign-in account. This is the preparation for adding scopes.
+  [self OAuthLoginWithOptions:nil
+                    authError:nil
+                   tokenError:nil
+      emmPasscodeInfoRequired:NO
+                keychainError:NO
+               restoredSignIn:YES
+               oldAccessToken:NO
+                  modalCancel:NO];
+
+  XCTAssertNotNil(_signIn.currentUser);
+
+
+  GIDSignInInternalOptions *options = [GIDSignInInternalOptions defaultOptionsWithConfiguration:nil
+                                                                      presentingViewController:nil
+                                                                                     loginHint:nil
+                                                                                 addScopesFlow:YES
+                                                                                      callback:nil];
+
+  id profile = OCMStrictClassMock([GIDProfileData class]);
+  OCMStub([profile email]).andReturn(kUserEmail);
+
+  OCMStub([_user authentication]).andReturn(_authentication);
+  OCMStub([_authentication clientID]).andReturn(kClientId);
+  OCMStub([_user serverClientID]).andReturn(nil);
+  OCMStub([_user hostedDomain]).andReturn(nil);
+
+  OCMStub([_user openIDRealm]).andReturn(kOpenIDRealm);
+  OCMStub([_user profile]).andReturn(profile);
+  OCMStub([_user grantedScopes]).andReturn(@[kGrantedScope]);
+
+  [self OAuthLoginWithOptions:options
+                    authError:nil
+                   tokenError:nil
+      emmPasscodeInfoRequired:NO
+                keychainError:NO
+               restoredSignIn:NO
+               oldAccessToken:NO
+                  modalCancel:NO];
+
+  NSArray<NSString *> *grantedScopes;
+  NSString *grantedScopeString = _savedAuthorizationRequest.scope;
+
+  if (grantedScopeString) {
+    grantedScopeString = [grantedScopeString stringByTrimmingCharactersInSet:
+        [NSCharacterSet whitespaceCharacterSet]];
+    // Tokenize with space as a delimiter.
+    NSMutableArray<NSString *> *parsedScopes =
+        [[grantedScopeString componentsSeparatedByString:@" "] mutableCopy];
+    // Remove empty strings.
+    [parsedScopes removeObject:@""];
+    grantedScopes = [parsedScopes copy];
+  }
+  
+  NSArray<NSString *> *expectedScopes = @[kNewScope, kGrantedScope];
+  XCTAssertEqualObjects(grantedScopes, expectedScopes);
+
+  [profile stopMocking];
 }
 
 - (void)testOpenIDRealm {
@@ -1023,10 +1079,7 @@ static void *kTestObserverContext = &kTestObserverContext;
     }
   } else {
     XCTestExpectation *expectation = [self expectationWithDescription:@"Callback called"];
-    [_signIn signInWithConfiguration:_configuration
-            presentingViewController:_presentingViewController
-                                hint:_hint
-                            callback:^(GIDGoogleUser * _Nullable user, NSError * _Nullable error) {
+    GIDSignInCallback callback = ^(GIDGoogleUser * _Nullable user, NSError * _Nullable error) {
       [expectation fulfill];
       if (!user) {
         XCTAssertNotNil(error, @"should have an error if user is nil");
@@ -1034,12 +1087,24 @@ static void *kTestObserverContext = &kTestObserverContext;
       XCTAssertFalse(self->_callbackCalled, @"callback already called");
       self->_callbackCalled = YES;
       self->_authError = error;
-    }];
+    };
+    if (options.addScopesFlow) {
+      [_signIn addScopes:@[kNewScope]
+        presentingViewController:_presentingViewController
+                        callback:callback];
+    } else {
+      [_signIn signInWithConfiguration:_configuration
+              presentingViewController:_presentingViewController
+                                  hint:_hint
+                              callback:callback];
+    }
 
     [_authorization verify];
     [_authState verify];
 
     XCTAssertNotNil(_savedAuthorizationRequest);
+    NSDictionary<NSString *, NSObject *> *params = _savedAuthorizationRequest.additionalParameters;
+    XCTAssertEqualObjects(params[@"include_granted_scopes"], @"true");
     XCTAssertNotNil(_savedAuthorizationCallback);
     XCTAssertEqual(_savedPresentingViewController, _presentingViewController);
 
@@ -1098,15 +1163,20 @@ static void *kTestObserverContext = &kTestObserverContext;
   [[[_authState expect] andReturn:tokenResponse] lastTokenResponse];
 
   // SaveAuthCallback
-  [[[_user stub] andReturn:_user] alloc];
   __block OIDAuthState *authState;
   __block GIDProfileData *profileData;
 
   if (keychainError) {
     _saveAuthorizationReturnValue = NO;
   } else {
-    (void)[[[_user expect] andReturn:_user] initWithAuthState:SAVE_TO_ARG_BLOCK(authState)
-                                                  profileData:SAVE_TO_ARG_BLOCK(profileData)];
+    if (options.addScopesFlow) {
+      [[_user expect] updateAuthState:SAVE_TO_ARG_BLOCK(authState)
+                          profileData:SAVE_TO_ARG_BLOCK(profileData)];
+    } else {
+      [[[_user stub] andReturn:_user] alloc];
+      (void)[[[_user expect] andReturn:_user] initWithAuthState:SAVE_TO_ARG_BLOCK(authState)
+                                                    profileData:SAVE_TO_ARG_BLOCK(profileData)];
+    }
   }
 
   if (restoredSignIn && !oldAccessToken) {
@@ -1139,8 +1209,12 @@ static void *kTestObserverContext = &kTestObserverContext;
   _keychainRemoved = NO;
   _keychainSaved = NO;
   _authError = nil;
-  [[[_user expect] andReturn:_authentication] authentication];
-  [[[_user expect] andReturn:_authentication] authentication];
+
+  if (!options.addScopesFlow) {
+    [[[_user expect] andReturn:_authentication] authentication];
+    [[[_user expect] andReturn:_authentication] authentication];
+  }
+
   __block GIDAuthenticationAction action;
   [[_authentication expect] doWithFreshTokens:SAVE_TO_ARG_BLOCK(action)];